fix: PR Review 반영 - club search api hook, create query 로직 분리 - Responsive Component 구현, UI와 로직 분리 - useDrawer open 함수 인자 obejct로 변경
