+++ src/components/home/Club/RecommendClub.tsx
@@ -1,6 +1,6 @@
 import { css } from '@styled-system/css'
 
-import { useGetClubRecommended } from '@/api/hooks/institution'
+import { useGetClubRecommended } from '@/api/hooks/club'
 import ClubProfile from '@/components/home/Club/ClubProfile'
 import ClubSectionTitle from '@/components/home/Club/ClubSectionTitle'
