+++ src/util/DrawerProvider.tsx
@@ -3,15 +3,19 @@ import { createContext, PropsWithChildren, ReactNode, useState } from 'react'
 import Drawer from '@/components/ui/drawer'
 import useScrollLock from '@/util/hooks/useScrollLock'
 
-export const DrawerContext = createContext<{
-  open: (element: ReactNode) => void
-  close: () => void
-} | null>(null)
-
+interface DrawerOpenProps {
+  element: ReactNode
+  option?: DrawerOpenOption
+}
 interface DrawerOpenOption {
   lockScroll?: boolean
   // TODO: Add Animation Opt or something
 }
+
+export const DrawerContext = createContext<{
+  open: (props: DrawerOpenProps) => void
+  close: () => void
+} | null>(null)
 
 export const DrawerProvider = ({ children }: PropsWithChildren) => {
   const [contents, setContents] = useState<ReactNode | null>(null)
@@ -19,8 +23,8 @@ export const DrawerProvider = ({ children }: PropsWithChildren) => {
 
   const { lockScroll: lock, unlockScroll: unlock } = useScrollLock()
 
-  const open = (element: ReactNode, option?: DrawerOpenOption) => {
-    const { lockScroll = true } = option ?? {}
+  const open = ({ element, option = {} }: DrawerOpenProps) => {
+    const { lockScroll = true } = option
     setContents(element)
     if (lockScroll) lock()
   }
